asset pipeline pseudo/rename stage + Str + language support

+++ src/main/java/com/vke/core/file/deflate/compress/lz77/Lz77Symbol.java
@@ -1,5 +1,5 @@
 package com.vke.core.file.deflate.compress.lz77;
 
-public record Lz77Symbol(boolean isLiteral, byte literal, int length, int distance) {
+public record Lz77Symbol(boolean isLiteral, int literal, int length, int distance) {
 
 }
